adaption with optional forwarding

+++ scripts/call_ncglue.sh
@@ -8,7 +8,7 @@ date
 SHRHOST=`hostname` 
 SHRHOST=${SHRHOST//rpi-/} # gets hostname (if neccessary cut 'rpi-')
 
-REMOTEHOST="pid@aero.nck.ggki.hu"
+REMOTEHOSTFILE="/home/pi/scripts/remotehost.txt"
 
 while true
 do
@@ -25,29 +25,31 @@ do
         ;;
       2) # exit code for finished job
         echo "job seems to be ready"
-        
-        if [ -s `pgrep -n rsync` ] ; then
-        
-            echo "rsyncing to $REMOTEHOST:$SHRHOST-data/"
-            OPWD=`pwd`
-            cd /home/pi/data
-            rsync -av --exclude tmp/ /home/pi/data/ $REMOTEHOST:$SHRHOST-data/
-            
-            if [ $? -eq 0 ]; then  #if successful, delete local files older than 3 days
-                find /home/pi/data/ -mindepth 3 -mtime +3 -type f -exec rm {} \;
-                echo "files deleted"
+        if [ -s $REMOTEHOSTFILE ] ; then
+            REMOTEHOST=`cat $REMOTEHOSTFILE`
+            if [ -s `pgrep -n rsync` ] ; then
+                echo "rsyncing to $REMOTEHOST:$SHRHOST-data/"
+                OPWD=`pwd`
+                cd /home/pi/data
+                rsync -av --exclude tmp/ /home/pi/data/ $REMOTEHOST:$SHRHOST-data/
+                if [ $? -eq 0 ]; then  #if successful, delete local files older than 3 days
+                    find /home/pi/data/ -mindepth 3 -mtime +3 -type f -exec rm {} \;
+                    echo "files deleted"
+                fi
+                cd $OPWD
+                exit 0
+            else
+                echo "rsync is already running. terminating"
+                exit 0
             fi
-            
-            cd $OPWD
-            exit 0
         else
-            echo "rsync is already running. terminating"
+            echo "no remote host defined."
             exit 0
         fi
         ;;
       *) #
         echo "ncglue: undefined exit code"
         exit 1
-        ;;        
+        ;;
     esac
 done
